Working draft [fast&dirty]

+++ Makefile
@@ -1,3 +1,6 @@
+# todo: refactor
+LAST_COMMIT_HASH = $(shell git rev-parse --verify HEAD)
+
 include dev/.env
 export
 export PATH := $(shell pwd)/tmp:$(PATH)
@@ -28,20 +31,32 @@ example: custom_ca
 ifdef CI # CI is set in Github Actions
 	cd test_example; SSL_CERT_FILE=${SSL_CERT_FILE} CURL_CA_BUNDLE=${CURL_CA_BUNDLE} vagrant up --provision
 else
-	cp -f docker/conf/certificates/*.crt test_example/docker/conf/certificates
+	if [ -f "docker/conf/certificates/*.crt" ]; then cp -f docker/conf/certificates/*.crt test_example/docker/conf/certificates; fi
 	cd test_example; SSL_CERT_FILE=${SSL_CERT_FILE} CURL_CA_BUNDLE=${CURL_CA_BUNDLE} CUSTOM_CA=${CUSTOM_CA} ANSIBLE_ARGS='--extra-vars "local_test=true"' vagrant up --provision
 endif
 
 # clean commands
-destroy-box:
+destroy:
 	vagrant destroy -f
+	rm terraform.tfstate || true
+	rm terraform.tfstate.backup || true
+	rm example/terraform.tfstate || true
+status:
+	vagrant global-status
+format:
+	./format.sh
 
 remove-tmp:
 	rm -rf ./tmp
 
-clean: destroy-box remove-tmp
+clean: destroy remove-tmp
 
 # helper commands
 update-box:
 	@SSL_CERT_FILE=${SSL_CERT_FILE} CURL_CA_BUNDLE=${CURL_CA_BUNDLE} vagrant box update || (echo '\n\nIf you get an SSL error you might be behind a transparent proxy. \nMore info https://github.com/fredrikhgrelland/vagrant-hashistack/blob/master/README.md#if-you-are-behind-a-transparent-proxy\n\n' && exit 2)
 
+proxy-to-minio:
+	consul connect proxy -service minio-local -upstream minio:9000 -log-level debug
+docker-local-build:
+	docker build docker -t fredrikhgrelland/hive:${LAST_COMMIT_HASH}
+	docker tag fredrikhgrelland/hive:${LAST_COMMIT_HASH} fredrikhgrelland/hive:test
